requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
argon2-cffi
PyJWT
pydantic[email]
